add build test and 5 tests

+++ src/ft_memcpy.c
@@ -6,18 +6,17 @@
 /*   By: angomes- <angomes-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/01 20:30:33 by angomes-          #+#    #+#             */
-/*   Updated: 2023/05/03 17:22:57 by angomes-         ###   ########.fr       */
+/*   Updated: 2023/05/05 18:10:56 by angomes-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
 #include <string.h>
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	char	*cpysrc;
 	char	*cpydest;
-	int		i;
+	size_t		i;
 
 	cpysrc = (char *)src;
 	cpydest = (char *)dest;
